requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
gevent==21.1.2
gunicorn==20.1.0
mysql-connector-python==8.0.24